refactor: Better Command Terminology (#7197)
