+++ packages/discord.js/src/index.js
@@ -73,7 +73,6 @@ exports.Application = require('./structures/interfaces/Application');
 exports.ApplicationCommand = require('./structures/ApplicationCommand');
 exports.AutocompleteInteraction = require('./structures/AutocompleteInteraction');
 exports.Base = require('./structures/Base');
-exports.BaseCommandInteraction = require('./structures/BaseCommandInteraction');
 exports.BaseGuild = require('./structures/BaseGuild');
 exports.BaseGuildEmoji = require('./structures/BaseGuildEmoji');
 exports.BaseGuildTextChannel = require('./structures/BaseGuildTextChannel');
@@ -82,13 +81,14 @@ exports.BaseMessageComponent = require('./structures/BaseMessageComponent');
 exports.ButtonInteraction = require('./structures/ButtonInteraction');
 exports.CategoryChannel = require('./structures/CategoryChannel');
 exports.Channel = require('./structures/Channel').Channel;
+exports.ChatInputCommandInteraction = require('./structures/ChatInputCommandInteraction');
 exports.ClientApplication = require('./structures/ClientApplication');
 exports.ClientPresence = require('./structures/ClientPresence');
 exports.ClientUser = require('./structures/ClientUser');
-exports.Collector = require('./structures/interfaces/Collector');
 exports.CommandInteraction = require('./structures/CommandInteraction');
+exports.Collector = require('./structures/interfaces/Collector');
 exports.CommandInteractionOptionResolver = require('./structures/CommandInteractionOptionResolver');
-exports.ContextMenuInteraction = require('./structures/ContextMenuInteraction');
+exports.ContextMenuCommandInteraction = require('./structures/ContextMenuCommandInteraction');
 exports.DMChannel = require('./structures/DMChannel');
 exports.Emoji = require('./structures/Emoji').Emoji;
 exports.Guild = require('./structures/Guild').Guild;
@@ -116,7 +116,7 @@ exports.MessageAttachment = require('./structures/MessageAttachment');
 exports.MessageButton = require('./structures/MessageButton');
 exports.MessageCollector = require('./structures/MessageCollector');
 exports.MessageComponentInteraction = require('./structures/MessageComponentInteraction');
-exports.MessageContextMenuInteraction = require('./structures/MessageContextMenuInteraction');
+exports.MessageContextMenuCommandInteraction = require('./structures/MessageContextMenuCommandInteraction');
 exports.MessageEmbed = require('./structures/MessageEmbed');
 exports.MessageMentions = require('./structures/MessageMentions');
 exports.MessagePayload = require('./structures/MessagePayload');
@@ -144,7 +144,7 @@ exports.ThreadChannel = require('./structures/ThreadChannel');
 exports.ThreadMember = require('./structures/ThreadMember');
 exports.Typing = require('./structures/Typing');
 exports.User = require('./structures/User');
-exports.UserContextMenuInteraction = require('./structures/UserContextMenuInteraction');
+exports.UserContextMenuCommandInteraction = require('./structures/UserContextMenuCommandInteraction');
 exports.VoiceChannel = require('./structures/VoiceChannel');
 exports.VoiceRegion = require('./structures/VoiceRegion');
 exports.VoiceState = require('./structures/VoiceState');
